LAB PROJECT/requirements.txt
numpy==1.24.3
Pillow==12.0.0
scikit-learn==1.7.2
joblib==1.5.2
scikit-image==0.21.0
